requirements.txt
aiogram>=3.4,<4
SQLAlchemy[asyncio]==2.0.*
asyncpg
alembic

redis>=5
apscheduler>=3.10

pydantic-settings>=2.4
python-dotenv>=1.0

google-api-python-client>=2.140
google-auth>=2.30
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.2

gspread>=6.0
httpx>=0.27
loguru>=0.7
